openssl: fix ARM build, add tons of TODOs
The OpenSSL build currently is absolutely broken for anything not x86.  x86_64
compilation happens to work as a result of invoking a best effort guess script
that wraps the actual configuration script.  This needs to be replaced with a
proper platform detection and configuration selection logic.  The mapping for
the various platforms to the configuration option has been added as a comment to
permit the next person to retain part of their sanity.

This at least permits builds to pass on ARM again.