profiles/options: provide decent defaults for phonon+dbusmenu
The qt5 options are only needed for KDE Frameworks 5 for now. So
default to the behaviour of the old exheres. This makes Jenkins
happy again (look at
https://galileo.mailstation.de/jenkins/job/kde/133/consoleFull for
example failure).
Other possibility to solve this would be to hard enable qt4 and only
provide a qt5 option. But I think the chosen approach is better because
it makes switching to the new Qt version easier.

Change-Id: I15fed6b3eb88e220efcb7427e90cb13ffdf8a3a2