postgresql: remove older (<9.2) versions
Versions <9.2 are failing tests for multiple reasons.
One timeouts on initdb, one suffers path length exaustion on initdb - both
caused by its inability to cope with the symlink /usr/share/zoneinfo/posix.
Issue is known by upstream and won't fix, but inexistent on PG>=9.2.
Also, 9.0 and 9.1 (tested locally only) bash against sydbox because of a
missing whitelist rule.

Only the last one is fixed, however it is still best to remove the
problematic versions along with the cruft in the exlib that was needed to
support them.

Change-Id: I62ef21b861d5a3a326425fc9086319c91b021ddd