postgresql: change some of the builtin modules
Some modules are extremely useful if/when the server halts, behaves
unexpectedly or an unusual administrative task appears (such as
triggering log rotation from pgadmin). When that occurs, rebuilding and
restarting the server to install them to debug/fix the issue would actually
destroy the relevant information (such as current locks and statements) we
wanted to analyze with the modules. Since they are small, it's best to
always have them built. There are four now, but others probably should
be added soon.

Also, pg_upgrade requires the binaries of both major versions at the
same time to the upgrade, so it is not really useful unless postgresql
were slotted - it was turned into an option instead.

Revbumped current versions.

Change-Id: I6b8c8657e267f00d0f07286613989d2f9ad91de0