ca-certificates-20140223-r1: Optionally re-add CAcert (default-enabled)
Debian, the package of which we're using, decided to remove CAcert's
certificates due to CAcert's failure to undergo security audits.
cf. https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=718434

This decision was controversial and, thus, I feel it's appropriate to
give people a choice here and add an option to install CAcert's
certificates. Since CAcert has been included for years, the option is
enabled in the base profile (in a separate commit).