python: don't add LIBDIR to LDFLAGS
This causes LDSHARED in _sysconfigdata.py to have –L/usr/${LIBDIR} added to
the end of it. This in turn causes extension modules being linked by distutils
to always search LIBDIR first. This is causing the python bindings for
libtorrent-rasterbar to be linked against the system libtorrent rather than the
one from the build. This line was added as part of multibuildifying python, see
6d90eb7d, but doesn’t seem to be necessary.

Change-Id: If74a11fec59d406170bed124f9c7290709efa077