adapt skeleton-filesystem-layout for multiarch
The main symlink is
    /usr/host -> ${CHOST}
that allows switching the "active" host with a single symlink.
It also provides compatibility with things that hardcode arch-dependent
paths in arch-independent paths. It's a bad idea but it happens anyway.
e.g. /etc/dbus-1/system.conf hardcoding /usr/host/libexec/dbus-daemon-launch-helper
or PATH in /etc/profile.

Rest of the layout:
    /bin  -> usr/host/bin
    /sbin -> usr/host/sbin
    /lib  -> usr/host/lib

    /usr/bin     -> host/bin
    /usr/sbin    -> host/sbin
    /usr/lib     -> host/lib
    /usr/libexec -> host/libexec
    /usr/include -> host/include

Should be compatible with normal /usr/local layout.
It can be argued that this would break some things? maybe?
    /usr/local -> host/local

Not sure if we need:
    /usr/src -> host/src

I'd say yes on the latter for convenience, as I build kernels there.

Questions?
Do we still need the LDPATH env file? I think not.

Change-Id: If284cbbf96cd59ed38d59b35fa59767405ffc8d3