postgresql: use libuuid instead of ossp in the near future
As of 9.4, the generator functions of the extension uuid-ossp may depend on
one of three implementations: ossp, util-linux's libuuid or bsd's libuuid.
Since both us and upstream agree that the ossp implementation is a
rotten mess, let's switch to util-linux ASAP. The extension name will
change in the future, but so far they kept it the same.

Change-Id: I72eb167eb71c846f06ca2a46de0d74dee8363a07