Restore gnupg-2.0.26
Several gnome packages still have hard dependencies on gnupg[=2.0*], so
restore this so their deps can still be satisfied. These applications
probably won't be updated until the next gnome release series.

GnuPG 2.0 is still a stable, supported release, I don't see any reason
to remove it completely for the time being.

This restores the exheres to the same state as it existed in
27277104a4b0d2878a087c815db31c7a460814f1

Change-Id: Ia82bb1067e66956399ebb4204c7eb28074c5422f