github.exlib: change tarball name
This prevents issues when moving from
    require github [ rev=93dcb28ff08181029862b7072cd56295a16704d3 ]
to
    require github [ tag=v${PV} ]

because in the first case the tarball contains
    ${PN}-93dcb28ff08181029862b7072cd56295a16704d3
and in the second case
    ${PN}-${PV}

so if the tarball is already downloaded, it will have to be removed.

switch from .tgz to .tar.gz to avoid that.
Also make the tarball name reflect what's inside by changing it
from ${PNV} to ${reponame}-${sanitized_tag/commit}

Change-Id: I3dbc924e1a0002364d076189e4d5ffae10a97166