postgresql: fix uuid configure flag
This change simplifies the logic quite a bit. For >=9.4, we can always
let ./configure look into e2fs since libuuid is in util-linux, which is
in the system set, and it will only ever link to libuuid for the uuid-ossp
module if it is enabled (and also there is no --without-uuid apparently).

Didn't revbump because for all compilations that worked before
(9.3+uuid, 9.3-uuid and 9.4+uuid), the output is exactly the same; and for
the one combination that didn't (9.4-uuid), now it compiles cleanly.

Change-Id: I43e57382337f8fe1e0b48750d1d83405a65a2e36