stages.conf: Add app-editors/emacs to the stages set
I was looking through our documentation and
http://www.exherbo.org/docs/gettingstarted.html links to outdated stages
whereas Jenkins is building fresh stages every day which are not
considered official due to them containing emacs.
(cf.
http://lists.exherbo.org/pipermail/exherbo-dev/2014-May/001318.html)

The install guide mentions both. I think it's finally time to end this
farce and solve it by having both emacs *and* vim in the stages set.

Change-Id: Ia5aa2c417b279a7077e9e5a9b8168ca1506d876c