gcc: make libgcc_s build work again
The build of libgcc_s would fail due to missing crt*.o objects in the search
path.  The configure script will use FLAGS_FOR_TARGET as part of the driver
invocation for the link.  This will have /usr/${target}/${target}/lib in the
search path, rather than /usr/${target}/lib.  Forcefully adjust this until we
figure out the root cause for the missing path in the search for this particular
link.  Fortunately, this only effects build time and not runtime behaviour for
the compiler.