cmake: Add doc option
It's reasonable to assume that most people using cmake mostly want
it as a build system for a package they want. Because Sphinx also
pulls in a few python deps, making it optional appears justified.

Also remove the unnecessary emacs option from the exlib because
elisp-optional.exlib already takes care of that.

Change-Id: Icbc93d7c84d7c400f0fd7fb9bd1d6eb66f6713b4