libgcc: reduce overhead of build
This is a very creative solution with a hack.  We now only build the
build-libiberty, libdecnumber, gengtypes and family for instruction headers.  We
modify the Makefile for just building libgcc.mvars.  This allows us to
cross-compile libgcc without a full build of gcc which is unnecessary.

The hack is due to libdir being incorrect at install time, resulting in the
library going to the wrong libdir.  Fixing that should permit us to avoid the
install time hack.

This is the last piece of the cross puzzle.  We now have a full toolchain in
static-gcc (which needs alternatives handling), and split runtimes which are
cross-compiled to the necessary targets.