static-gcc: add threads option; go frontend
The go frontend does not require dependencies; enable it unconditionally.

Add the threads option.  This option allows us to effectively use the same
package for the "bootstrap" and normal usage of gcc.  Unfortunately, packages
rely on the output of `$(CC) -v` to determine the threading model, which is
reported as single if we don't explicitly enable the threading model.  This is
unfortunate, but enables the correct building of libstc++, which is important.