Revert "migration-gcc: be more clever for C++"
Apparently, we have no choice but to build and rebuild the runtimes as we need
to use them to build the compiler (which was not necessary previously since we
built them together, resulting in the compiler building and using the version
that we were building rather than a merged version).  All that trickery for
naught.