libstdc++: restore workaround
Unfortunately, we need to setup the gthr-default symlink as configure pushes the
path to libgcc via TOPLEVEL_FLAGS into the test.  The result is that the
system's installed image is not consulted for gthreads availability (due to the
bundled build).  This broke when we split apart the runtimes.

Make the check more clever and use the same check as the configure script to
setup the link to permit building with single and posix threading models.