static-gcc: do not inhibit libc when building threaded
Since supporting the threading model requires the presence of libc, when
building with threads, we can assume that we can build against the target libc.

This is particularly important for the static libgcc which is built as part of
the build.  On Linux x86_64, the inhibition of libc will result in libgcc being
built without support eh_frame_hdr.  This is problematic as the compiler will
enable the use of the eh_frame_hdr LSB extension.  The result is that the unwind
FDE will be ignored and the landing pad cannot be detected for exceptions.  As a
result exception handling is impossible.