gcc: correct C++ handling for non-native targets, cleanup
Specify the C++ include directory for all targets.  When building the
cross-compiler, the default path uses an alternate path which does not resolve
to the desired layout (it assumes a mixed host/target environment rather than
the prefix-root approach).  This should fix cross-compilation of C++ code.

Cleanup some of the libpath handling as we no longer have /lib due to
/usr-merge.  This allows for a single location for all libraries which is much
easier to work with.  No need to provide the joined path.

Drop the local variable containing the libpath in src_install as that is no
longer referenced.