sys-devel/gcc: fix catching of alternatives
Disable expansion of globs for alternatives if there is no match.

By default nullglob is off and if bash cannot expand pattern it leaves
it as is and we store string as pattern with prefix '/'. Later we
expand that string without quotes and bash perform glob expanding. But
now it is successful since there is paths that match that pattern.

Thanks to Benedikt Morbach for guidance.

Change-Id: I2da90460e7d6365a1ff409644c5650d9ee29512f