dev-libs/boost: specifically use python2, don't use eclectic symlink
Using `python` can cause issues if your default `python` provider
is set to 3 in `eclectic`; so, just use `python2`.

Also, TODO for the future: we should support python3 for boost compilation
since boost supports python 3 just fine and we don't have a reason not to.
I would do it myself but I wouldn't trust myself to do it well being that
I'm not very experienced with boost or python packages.

Change-Id: I6957e659a6437623fbe0df430685b695612f2c6f