libgcc: ensure that the right *flags are used
This is particularly important when not building a multilib.  In order to reduce
the build overhead, we simply build libiberty, libdecnumber, and parts of gcc.
However, the builds for these is interesting.  We need to build libiberty for
the *BUILD*, as well as gcc (as gcc is cross-compiled).  We then use that
configuration information to build libgcc.  Because of the variable handling, we
were accidentally bleeding the target flags into the build flags, resulting in
build failures if there were flags in use on the target that did not apply to
the host.  This would silently succeed on multilib targets, since the flags
would remain applicable.  We not more stringently filter the variables.

Tested with a cross-compilation of libgcc from x86_64-pc-linux-gnu to
powerpc64-unknown-linux-gnu.