vim: use tarballs from git
We currently use the release tarball (7.4) and add the patches that
upstream releases on top (.640)

These patches only ever touch the vim code and not the runtime. So the
syntax files and other runtime things I don't care about haven't
recieved any changes since 7.4 was release (2013).

Upstream says:
    "After you obtained the archive and applied a bazillion
     patches, you still don't have the latest version of the runtime files.
     Because whatever.
     If you can't wait for the next Vim release to get the latest runtime
     files, this is the way to get them:
             <obscure semi package manager thing>
             or rsync
     Because YOLO!"
     (slightly paraphrased from http://www.vim.org/runtime.php)

So getting an up-to-date runtime is totally supported afaict.

However, both code and rumtime are updated in the same upstream repo all
the time[1] and the patches are just cherry-picks of the commits that
touch the code.

So instead of getting a tarball + 700 patches + another hand-rolled
tarball for the runtime or getting it from scm like upstream suggests,
let's just use the tags from which the patches are made in the first
place, but with the rest being up-to-date as well.

[1] You ask why they can't be released from that in one tarball or maybe two?
    ...
    I don't know. Maybe they think it's more fun for us this way.
    *checks neovim status again*

Change-Id: I29353ffc6bebdc95eaf97d1eb6e7962c23ae448a