icu: fix cross compilation to different libcs of the same architecture
ICU's default INVOKE overrides LD_LIBRARY_PATH pointing to the freshly
built libraries, and its default icucross.mk is supposed to override it
in a cross-friendly manner, but our icucross.mk didn't. If the host was
built with a different libc than the target, but the arches match the
loader actually tries to load the freshly-built libraries but fails.
Without matching arch the loader just skipped the new libraries because
the architecture didn't match the one of the binary.

Change-Id: I509df77e524d83bf7fbd6bb953684624dd4a831d