openssh: Follow up to 404f962fa2680efe2803663ed562
I've kind of lost track of the original patch on gerrit before it was
merged. Although it was an improvement (the weak keys will not be
created if you have the strong ones) the weak host keys are still
generated if one doesn't have any key at all.
This patch introduces a bash script replacing the ssh-keygen -A
invocation to fix this.

Change-Id: I93acd6e80e8e9a27740c24e8646d5690c041f2d4