gcc: drop silly die
this doesn't even allow e.g. powerpc, which is in gcc's targets...
It only protects against adding another target (say armv8) and
getting the wrong paths.

But it doesn't protect you against adding armv7-whatever-linux-gnueabi,
so it's useless anyway.

Replace it with a comment.

We have to change the patch at some point to hook into the buildsystem
if we want to support different paths for armv8 or vendor thingies, so
we can have the build die from within there when we do that.

Change-Id: I5bf5e8f9cc286427f48af1c7287c4ebb4dfb1e3c