less: Really fix LICENCES
192f0e2a2eaa46598e9d7f6ec7b774cb77dbb063 changed LICENCES from GPL-3 to
BSD-2. Turns out neither is completely right. First, there is not
really a difference between less from gnu.org and greenwoodsoftware.com
other than that the latter hosts beta versions. Furthermore less can be
used under the conditions of the GPL-3 OR BSD-2. The README clearly
states:

"This program is free software.  You may redistribute it and/or
    modify it under the terms of either:
    [...]
    1) The GNU General Public License, [...] version 3
    2) The Less License [...]"

The latter is a BSD-2 license.

Change-Id: I3bb9467915f78411e06439c1ea95b3417f32b073