pinentry: fix library config lookup on cross
pinentry dependencies libassuan and libgpg-error each provide a
config script for looking up installation details. When compiled
for cross, the native version is selected from the PATH, leading
to wrong headers being used.
This is fixed by using the configure script's --with-*-prefix=
switches to specify the correct target path.

Change-Id: I3a931227d8539879b9a9d2f80d5eb08429a5d349
Reviewed-on: https://galileo.mailstation.de/gerrit/4035
Reviewed-by: Jenkins <philantrop@gmail.com>
Reviewed-by: Kylie McClain <somasis@exherbo.org>