libgcc: pass the same --with-sysroot as we do for gcc
Without this inhibit_libc is set when cross-compiling, which causes some
code not to be built. I noticed this when cross-compiling from musl to
glibc of the same architecture causing code using backtrace(3) function
failing with SIGABRT, but mysteriously working when loading a
libgcc_s.so from a host where it wasn't cross-compiled.

Change-Id: Ic03a82b3642a0001d2826efc5b1960dd83ee43df
Reviewed-on: https://galileo.mailstation.de/gerrit/4160
Reviewed-by: Jenkins <philantrop@gmail.com>
Reviewed-by: Marc-Antoine Perennou <Marc-Antoine@Perennou.com>
Reviewed-by: Kylie McClain <somasis@exherbo.org>