mysql: version bump to 5.7.9 (masked)
This version ships its own systemd setup which will automatically create
an initial database on startup. During this process a temporary password
is assigned and can be viewed in the systemd journal.

Only include mysql in the virtual/mysql for now since I lack knowledge
of the other ones. Feel free to update the virtual.

Change-Id: I90ef051f1400f016030d1cc230ac0bac50b772e2
Reviewed-on: https://galileo.mailstation.de/gerrit/4392
Reviewed-by: Timo Gurr <tgurr@exherbo.org>