Mirrors: add osdn.jp
osdn.jp is the new name of sourceforge.jp (no relation to
sourceforge.net). It has a new set of mirrors, see
https://osdn.jp/docs/Mirrors (the mirror URLs were extracted from the
download page).

It's probably best to leave the sourceforge.jp mirrors for the time
being, since they'll continue to work for releases prior to the name
change; they can be removed once packages that reference them are
updated.

Change-Id: Id791293eda2f22fed6c4a74636e1c17a7e5ca1c1
Reviewed-on: https://galileo.mailstation.de/gerrit/4631
Reviewed-by: Jenkins <wk@mailstation.de>
Reviewed-by: Kylie McClain <somasis@exherbo.org>