setup-py.exlib: use --force
When installing packages with executables without --force setup.py
refuses to overwrite these executables when using multiple abis.

This means when using python_abis: 2.7 3.4 that the executable will end
up with a shebang of python 2.7, it would make more sense if we where to
pick the newest working version instead of the oldest.

I'm not sure if this change would require a newsitem to rebuild all
python files, because technically they will work just fine without a
rebuild, but it will change the contents of installed executables
mostlikely.

Also this change might be 'dangerous' in that it can expose problems
because executes where simply never ever run with anything but the
lowest python on people their systems. So it would require atleast a few
people to rebuild everything and verify that nothing crucial breaks.

Change-Id: I078c680d0ec8b95910a8c4626946728ea93c1e6e
Reviewed-on: https://galileo.mailstation.de/gerrit/4826
Reviewed-by: Marc-Antoine Perennou <Marc-Antoine@Perennou.com>
Reviewed-by: Wouter van Kesteren <woutershep@gmail.com>
Reviewed-by: Kylie McClain <somasis@exherbo.org>