go: restrict strip
go has this really strange issue where if you run dwz on its binaries,
it will somehow manage to mess up the binaries, which can result in
segfaults, or just corrupted binaries.

I'd prefer for there to be a way to just disable dwarf_compress for it;
things like symbols=compress and symbols=strip work just fine, dwarf_compress
is the only one that messes it up.

Change-Id: I2579343ab766f16492f24b31ba82e39e2ebd1433