libressl: install a c_rehash script for ca-certificates
Just calling openssl certhash /etc/ssl/certs has proven not to be enough
in ca-certificates, as /etc/ssl/certs/ca-certificates.crt is not updated
with that command in contrast to update-ca-certificates. So now we
provide a c_rehash binary that is compatible enough with openssl's so we
can use update-ca-certificates with it again.

Update-ca-certificates is called here in libressl so the certificate
store is rebuilt if upgrading from a broken version. The alternative to
this would have been to revbump ca-certificates after the fix and make
it depend on the fixed revision of libressl. This isn't feasible because
we have multiple versions the fix was applied to and no way to specify
this dependency in ca-certificates.

Change-Id: Id54da5f72b5534de06a37ae5ebaebfcb9cb57676
Reviewed-on: https://galileo.mailstation.de/gerrit/5013
Reviewed-by: Marc-Antoine Perennou <Marc-Antoine@Perennou.com>
Reviewed-by: Kylie McClain <somasis@exherbo.org>