Revert "libgpg-error: fix awful build system with musl cross"
This reverts commit a5d670d831c35132c63d6817fe90c2b9035d1f8b.

This definition does _NOT_ work for all architectures. Don't use it
unconditionally.

If this is the wrong architecture-specific object libgpg-error fails
with "posix-lock.c:119: get_lock_object: Assertion `!"sizeof lock obj"'
failed.".

This assert is triggered if "(sizeof (gpgrt_lock_t) < sizeof
(_gpgrt_lock_t))". The size of _gpgrt_lock_t (defined in
src/posix-lock-obj.h) is on all supported platforms dependent on the
libc's pthread_mutex_t object. It's not possible to make this work in
the general case without generating the header specifically for the
targeted libc ABI.

The right way to fix this is to generate the lock objects on the target
system and include them here for each system for which upstream doesn't
already ship them.

Change-Id: I8d5fb598dbbdbffcf2fc37c946bbb71e1691bbaa
Reviewed-on: https://galileo.mailstation.de/gerrit/5117
Reviewed-by: Jenkins <wk@mailstation.de>
Reviewed-by: Heiko Becker <heirecka@exherbo.org>
Reviewed-by: Kylie McClain <somasis@exherbo.org>