github.exlib: Allow pn to be different from repository name
This makes the exlib easier to use for cases where multiple packages are
hosted from a single generically-named github repository, and also
better handles autogenerated tarballs from poorly named repositories.

The change is backwards compatible with current usage of the exlib - a
new exparam is added to override the github repo name ("project") in
cases where it doesn't match the package name ("pn").

Change-Id: Ib3a8ba5e607fc505cff2fa15f6837e866218503b
Reviewed-on: https://galileo.mailstation.de/gerrit/5173
Reviewed-by: Jenkins <wk@mailstation.de>
Reviewed-by: Volodymyr Medvid <vmedvid@riseup.net>
Reviewed-by: Kylie McClain <somasis@exherbo.org>