cmake.exlib: set CMAKE_FIND_ROOT_PATH
"This variable is most useful when cross-compiling. CMake uses the paths
in this list as alternative roots to find filesystem items with
find_package(), find_library() etc."
(https://cmake.org/cmake/help/v3.4/variable/CMAKE_FIND_ROOT_PATH.html)

This fixes cross-compiling SDL:2, as it tries to use the libX11.so from
/usr/lib otherwise, failing a configure check (HAVE_XGENERICEVENT) that
actually needs to succeed for it to compile (resulting in the following
error: "conflicting types for 'XGenericEventCookie'").

CMAKE_SYSTEM_PREFIX_PATH is not enough as it would have been tried last
after libX11.so is already found in /usr/lib. CMAKE_PREFIX_PATH is
intended to be set by the project, which leaves us with
CMAKE_FIND_ROOT_PATH. For reasons yet to be determined this causes some
issues with building packages natively though, so we only pass when
we're actually cross compiling.

Change-Id: I9a1064d6c18169d520261ad295c89673f1d1e268
Reviewed-on: https://galileo.mailstation.de/gerrit/5012
Reviewed-by: Jenkins <wk@mailstation.de>
Reviewed-by: Heiko Becker <heirecka@exherbo.org>