binutils: disable STT_COMMON for everyone
While debugging the problem on musl it turned out that the problem is
not caused by musl not supporting STT_COMMON - they actually support it
to the same extend as glibc does. It's caused by a bug in binutils which
is not specific to musl, but which breaks musl's libc.so. The bug makes
bfd not emit relocations in some cases where it should. This'll be fixd
for now by just disabling STT_COMMON support for everyone until the bug
is fixed. It can be reenabled without problems when the bug is fixed.

Change-Id: I20a9ea7449317646e295471fbf6c7cc2a61255ee
Reviewed-on: https://galileo.mailstation.de/gerrit/5248
Reviewed-by: Jenkins <wk@mailstation.de>
Reviewed-by: Heiko Becker <heirecka@exherbo.org>