bash: revamp color handling in bashrc
The previous code stopped working with the release of coreutils 8.24.

This is mostly copied from Gentoo's bashrc. Here is their commit message
regarding the change.

"app-shells/bash: bashrc: drop custom parsing of dircolors databases #572582

Starting with coreutils-8.24, the dircolors TERM entries are run
through fnmatch rather than being a plain text string. This means our
parsing logic no longer works because we assumed fixed strings. It isn't
easy to process a list of path globs in bash, so rework the code to
always run the dircolors tool. We were doing this anyways in the
majority of cases, so it's not like we're adding that much overhead.
The only people who are negatively impacted are interactive colorless
terminals."

[ Gentoo commit 747857408315deb4b6a0ddd2a34484ca8e5ec2ff ]

Change-Id: I9fd659165722a5ca5a165e33d0cda19b96fb8b4b
Reviewed-on: https://galileo.mailstation.de/gerrit/5446
Reviewed-by: Jenkins <wk@mailstation.de>
Reviewed-by: Markus Rothe <markusr815@gmail.com>