Move all CFLAGS and CXXFLAGS to the global make.defaults
This allows for a few things. First, users don't have to
set CFLAGS for all their targets if they don't care what
CFLAGS they compile with, and just want to use -march=native.
Second, this'll allow for fixing the stages script by including
the CFLAGS/CXXFLAGS for the host target and cross target in
/etc/paludis/bashrc, without really messy string mutation done
by the stages script itself. The stages script will only need
to contain generic flags to be used for each target, and we can
grab the native flags for the target from arbor's make.defaults.

Change-Id: Idf4dd9ad938bfed0d7d1ff6cf68f8b403176f0ec
Reviewed-on: https://galileo.mailstation.de/gerrit/5562
Reviewed-by: Jenkins <wk@mailstation.de>
Reviewed-by: Bo Ørsted Andresen <zlin@exherbo.org>
Reviewed-by: Kylie McClain <somasis@exherbo.org>