gcc: don't provide c++
is provided by eclectic-gcc

I think the upgrade path is fine with the eclectic call.

Can't use blockers to rebuild eclectic-gcc after gcc afaict because if you use
e.g. gcc-4.9 and it builds
	gcc 4.9
	gcc 5.1
	eclectic-gcc
it will probably break at gcc-5.1 because ${target}-c++ is missing.
(didn't actually try)

And we can't know which gcc is in use at resolve time

Change-Id: I8ef34c51dfe39f8cd73cc8f3db1f21e043044f9b
Reviewed-on: https://galileo.mailstation.de/gerrit/5610
Reviewed-by: Wouter van Kesteren <woutershep@gmail.com>
Reviewed-by: Jenkins <wk@mailstation.de>
Reviewed-by: Benedikt Morbach <moben@exherbo.org>