emacs: 24.5 -> 25.1.50
A snapshot of the latest git commit is used, because it contains some
musl-related fixes, and they have not released a new pretest version yet.
The fixes are really invasive and deal with bootstrapping the lisp runtime,
and can't easily be backported to 24.5, or 25.0.92 (the last pretest version),
for that matter. 25.1.50 is what the version as of this git snapshot is.

Change-Id: Ifc13d3a3cbfe136de775ccbbcba4a0928515d590
Reviewed-on: https://galileo.mailstation.de/gerrit/5658
Reviewed-by: Wulf C. Krueger <philantrop@exherbo.org>