elinks: 0.12pre6, initial exheres
The latest stable release is 0.11.7; it's from 2011, and other distribution
use this pre-release version too; it's safe to say that it's stable enough.

It was already packaged in ::alip, but this is rewritten. the original was
only for scm.

No copyright assigned to alip because I didn't use his exheres for any
reference. (I just knew it needed to be rewritten since it was very old)

Change-Id: If525e2d957893ef80105edc126f40f41f5328b3e