go: 1.6 -> 1.6.2
In addition, we now use the official binaries from golang.org for
bootstrapping, rather than using the last C version of Go.
This appears to fix building on musl, and additionally means that
building Go is a lot faster, and gets rid of any future maintainence
overhead which using the last C version would entail (patching for
new GCC versions, etc.)

The binaries that we use are statically compiled and don't appear to
really depend on any libc specific features, so we can safely use them
on both glibc and musl without any issues.

Change-Id: I612a58bb53376073501d8c09816b2082efcbbbaf
Reviewed-on: https://galileo.mailstation.de/gerrit/5921
Reviewed-by: Jenkins <wk@mailstation.de>
Reviewed-by: Kylie McClain <somasis@exherbo.org>