linux-headers-4.6: fix musl usage
These patches are not upstream yet because I am waiting for someone more experienced
with the kernel internals to fix it. Additionally there has been talk of fixing this
in musl libc itself; as such, I am not going to submit these upstream, but in order
to prevent impact upon people not using musl, these patches will only be applied if
linux-headers is being installed for a musl target.

Change-Id: I5142aa398ea36d161d3a17494f1baa7800852701