freedesktop-mime: only update mime-database if it has changed
update-mime-database has a -n switch for use in pre/post install
scripts, which will ensure it only is updated if a file in
mime-dir/packages or a file within it is newer than mime-dir/version.
using this switch will save time being wasted when upgrading a package
that uses freedesktop-mime, since it will only update if a file in the
mimedir actually changes.

Change-Id: If0050619bca23f20defcd1a8b4d4885e2b322f58