gcc: repair gcc:4.9 build
The previous fix was indeed correct and fixed the issue.  However, the patch was
against gcc:6.  Construct a patch for gcc:4.9 and repair the build for that.
Make the patch application uniform by keeping each version's patchset local to
the version.  NFC for gcc:5.1, gcc:6, so no revbump for those.  Revbump gcc:4.9
for the new plugin support.

Change-Id: I5c79702c76e0e9e86d80be64abba1731ce7226d3
Reviewed-on: https://galileo.mailstation.de/gerrit/6472
Reviewed-by: Benedikt Morbach <moben@exherbo.org>