cmake.exlib: Make building reliably work with gcc:6
A change in gcc-6 made using -isystem with the default include
directory fail. A bug report [1] about this has been closed with
RESOLVED WONTFIX.
This problem became mostly apparent with cmake and for unknown
reasons even more so with cmake-3.6.x (which is why that version is
currently masked). I found out that this is caused by cmake not
filtering out our default include path /usr/host/include, which
differs from the usual /usr/include. This is tracked upstream at
[2]. Until this is fixed we can specify the implicit include dir
manually to safely allow building packages with the new gcc and
unmask the latest cmake.
Tested with gcc[=6.2.0], cmake[=3.6.2] and ~350 native cmake-based
packages and a few cross-compiled ones.

[1] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=70129
[2] https://gitlab.kitware.com/cmake/cmake/issues/16291

Change-Id: I68a7304d7cfd7c88eb61f46eab7fd228a262055b